minor bugfixes

+++ src/FFTLib/FFTManager.java
@@ -5,6 +5,7 @@ import fftlib.gui.FFTFailState;
 import fftlib.gui.InteractiveFFTState;
 import javafx.scene.Node;
 import javafx.scene.input.DataFormat;
+import tictactoe.misc.Database;
 
 import java.io.BufferedWriter;
 import java.io.File;
@@ -12,10 +13,10 @@ import java.io.FileWriter;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Paths;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
+import java.util.*;
 import java.util.function.BiFunction;
+
+import static misc.Config.FFT_OVERWRITE;
 
 
 public class FFTManager {
@@ -29,7 +30,8 @@ public class FFTManager {
     public FFT currFFT;
     private static FFTFailState failState;
     public static InteractiveFFTState interactiveState;
-    static BiFunction<Action, Integer, FFTMove> actionToMove;
+    public static BiFunction<Action, Integer, FFTMove> actionToMove;
+    public static BiFunction<HashSet<Literal>, Integer, FFTState> preconsToState;
 
     public static final DataFormat SERIALIZED_MIME_TYPE = new DataFormat("application/x-java-serialized-object");
 
@@ -45,6 +47,7 @@ public class FFTManager {
         gameBoardHeight = dim[0];
         gameBoardWidth = dim[1];
         actionToMove = gameSpecifics::actionToMove;
+        preconsToState = gameSpecifics::preconsToState;
         failState = gameSpecifics.getFailState();
         interactiveState = gameSpecifics.getInteractiveState();
 
@@ -56,6 +59,8 @@ public class FFTManager {
     }
 
     public static void save() {
+        if (!FFT_OVERWRITE)
+            return;
         BufferedWriter writer;
         try {
             writer = new BufferedWriter(new FileWriter(path));
@@ -147,6 +152,10 @@ public class FFTManager {
         return failState.getFailState(ps, nonLosingPlays);
     }
 
+    public FFT AutoGenFFT() {
+
+        return null;
+    }
 
 }
